14:00

+++ 7Blessing/Assets/Scripts/PowerUp.cs
@@ -8,7 +8,7 @@ public class PowerUp : MonoBehaviour {
     private int normalSpeed = 5;
     private int superJumpPower = 10;
     private int normalJumpPower = 7;
-    private int healingPower = 25;
+    private float healingPower = 25;
     private float duration = 5f;
     private GameObject pcikUpEffect;
     [SerializeField] AudioClip clip;
